// src/generated/api/models/CommentResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 모든 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CommentResponse
 */
export interface CommentResponse {
    /**
     * 
     * @type {number}
     * @memberof CommentResponse
     */
    commentId?: number;
    /**
     * 
     * @type {string}
     * @memberof CommentResponse
     */
    writer?: string;
    /**
     * 
     * @type {string}
     * @memberof CommentResponse
     */
    writerProfileImageUrl?: string;
    /**
     * 
     * @type {string}
     * @memberof CommentResponse
     */
    content?: string;
    /**
     * 
     * @type {Date}
     * @memberof CommentResponse
     */
    createdAt?: Date;
}

/**
 * Check if a given object implements the CommentResponse interface.
 */
export function instanceOfCommentResponse(value: object): value is CommentResponse {
    return true;
}

export function CommentResponseFromJSON(json: any): CommentResponse {
    return CommentResponseFromJSONTyped(json, false);
}

export function CommentResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CommentResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'commentId': json['commentId'] == null ? undefined : json['commentId'],
        'writer': json['writer'] == null ? undefined : json['writer'],
        'writerProfileImageUrl': json['writerProfileImageUrl'] == null ? undefined : json['writerProfileImageUrl'],
        'content': json['content'] == null ? undefined : json['content'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
    };
}

export function CommentResponseToJSON(json: any): CommentResponse {
    return CommentResponseToJSONTyped(json, false);
}

export function CommentResponseToJSONTyped(value?: CommentResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'commentId': value['commentId'],
        'writer': value['writer'],
        'writerProfileImageUrl': value['writerProfileImageUrl'],
        'content': value['content'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
    };
}